refac: #115 update project arch
- Simplify project architecture with a feature-based approach
