--- app/api/v1/endpoints/__init__.py
@@ -1,4 +1,0 @@
-from .token import router as token_router
-from .user import router as user_router
-
-__all__ = ["token_router", "user_router"]
