+++ app/domains/__init__.py
@@ -1,0 +1,4 @@
+from app.domains.auth.router import router as auth_router
+from app.domains.users.router import router as user_router
+
+__all__ = ["auth_router", "user_router"]
