+++ app/domains/auth/router.py
@@ -2,10 +2,10 @@ from datetime import timedelta
 
 from fastapi import APIRouter
 
-from app.api.v1.schemas.token import TokenRequest, TokenResponse
 from app.core.security import create_access_token
 from app.database.session import DbSession
-from app.dependencies.auth import authenticate_user
+from app.domains.auth.dependencies import authenticate_user
+from app.domains.auth.schemas import TokenRequest, TokenResponse
 from app.exceptions import UnauthorizedError
 
 ACCESS_TOKEN_EXPIRE_MINUTES = 30
