+++ app/domains/users/crud.py
@@ -1,8 +1,8 @@
 from sqlalchemy.orm import Session
 
 from app.core.security import hash_password
+from app.domains.users.schemas import UserCreate
 from app.models import User
-from app.operations.user.schemas import UserCreate
 
 
 def get_all_users(db: Session) -> list[User]:
