--- app/operations/user/schemas.py
@@ -1,27 +1,0 @@
-from pydantic import BaseModel, EmailStr
-
-from app.fields import PasswordStr
-
-
-class UserBase(BaseModel):
-    email: EmailStr
-    name: str
-    last_name: str
-
-
-class UserCreateBase(UserBase):
-    password: PasswordStr
-
-
-class UserCreate(UserCreateBase):
-    role_id: int
-    is_active: bool = True
-
-
-class UserUpdate(BaseModel):
-    email: EmailStr | None = None
-    password: PasswordStr | None = None
-    name: str | None = None
-    last_name: str | None = None
-    role_id: int | None = None
-    is_active: bool | None = None
