+++ tests/app/api/v1/endpoints/test_token.py
@@ -1,7 +1,7 @@
 from fastapi import status
 from fastapi.testclient import TestClient
 
-from app.api.v1.schemas.token import TokenResponse
+from app.domains.auth.schemas import TokenResponse
 from tests.factories.user import UserFactory
