+++ tests/app/dependencies/test_auth.py
@@ -3,7 +3,11 @@ from pydantic import SecretStr
 from sqlalchemy.orm import Session
 
 from app.core.security import create_access_token
-from app.dependencies.auth import authenticate_user, get_current_active_user, get_current_user
+from app.domains.auth.dependencies import (
+    authenticate_user,
+    get_current_active_user,
+    get_current_user,
+)
 from app.exceptions import NotFoundError, UnauthorizedError
 from tests.factories.user import UserFactory
